+++ test/Structuring.t.sol
@@ -5,6 +5,7 @@ import {Test} from "forge-std/Test.sol";
 import {MandateRegistry} from "../src/MandateRegistry.sol";
 import {AnchorLog} from "../src/AnchorLog.sol";
 import {Bond} from "../src/Bond.sol";
+import {SpendMeter} from "../src/SpendMeter.sol";
 import {Receipts} from "../src/Receipts.sol";
 import {IFdcVerification} from "@flarenetwork/flare-periphery-contracts/coston2/IFdcVerification.sol";
 import {IEVMTransaction} from "@flarenetwork/flare-periphery-contracts/coston2/IEVMTransaction.sol";
@@ -26,6 +27,7 @@ contract StructuringTest is Test {
     MandateRegistry reg;
     AnchorLog anchorLog;
     Bond bond;
+    SpendMeter meter;
     MockFdcEvm mock;
 
     address principal = makeAddr("principal");
@@ -47,7 +49,8 @@ contract StructuringTest is Test {
         reg = new MandateRegistry();
         anchorLog = new AnchorLog(reg);
         mock = new MockFdcEvm();
-        bond = new Bond(reg, anchorLog, IFdcVerification(address(mock)), 24 hours, 1 hours);
+        meter = new SpendMeter(reg);
+        bond = new Bond(reg, anchorLog, IFdcVerification(address(mock)), 24 hours, 1 hours, meter);
         reg.setBond(address(bond));
         vm.warp(1_800_000_000);